Apply Subheader component to Members page
