+++ app/components/members/index_page_header_component.html.erb
@@ -2,26 +2,5 @@
   render(Primer::OpenProject::PageHeader.new) do |header|
     header.with_title { page_title }
     header.with_breadcrumbs(breadcrumb_items, selected_item_font_weight: current_breadcrumb_element == page_title ? :bold : :normal)
-
-    header.with_action_button(scheme: :primary,
-                              mobile_icon: :plus,
-                              mobile_label: t('activerecord.models.member'),
-                              size: :medium,
-                              aria: { label: I18n.t(:button_add_member) },
-                              title: I18n.t(:button_add_member),
-                              id: "add-member-button",
-                              data: add_button_data_attributes) do |button|
-      button.with_leading_visual_icon(icon: :plus)
-      t('activerecord.models.member')
-    end
-
-    header.with_action_icon_button(mobile_icon: "filter",
-                                   scheme: :default,
-                                   icon: "filter",
-                                   label: I18n.t(:description_filter),
-                                   id: "filter-member-button",
-                                   aria: { label: I18n.t(:description_filter) },
-                                   class: "toggle-member-filter-link",
-                                   data: filter_button_data_attributes)
   end
 %>
